+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.0](https://github.com/alteral/xcmonkey/releases/tag/1.3.0)
+
+_January 29, 2023_
+
+### 🔄 Changed
+
+- `duration` option is renamed to `event-count`
+
+### ✅ Added
+
+- New test options that allow to
+  - specify gestures to exclude from the test
+  - set up throttle between events
+  - ignore app crashes
+
 ## [1.2.0](https://github.com/alteral/xcmonkey/releases/tag/1.2.0)
 
 _January 23, 2023_